+++ tests/testthat/test-eq5d3l.R
@@ -128,6 +128,14 @@ test_that("EQ-5D-3L Netherlands TTO gives correct answer", {
   expect_equal(eq5d3l(c(MO=3,SC=3,UA=3,PD=3,AD=3), "TTO", "Netherlands"), -0.329)
 })
 
+test_that("EQ-5D-3L Pakistan TTO gives correct answer", {
+  expect_equal(eq5d3l(c(MO=1,SC=1,UA=1,PD=1,AD=1), "TTO", "Pakistan"), 1)
+  expect_equal(eq5d3l(c(MO=2,SC=2,UA=2,PD=2,AD=2), "TTO", "Pakistan"), 0.783)
+  expect_equal(eq5d3l(c(MO=3,SC=2,UA=1,PD=2,AD=3), "TTO", "Pakistan"), 0.448)
+  expect_equal(eq5d3l(c(MO=1,SC=2,UA=3,PD=1,AD=2), "TTO", "Pakistan"), 0.627)
+  expect_equal(eq5d3l(c(MO=3,SC=3,UA=3,PD=3,AD=3), "TTO", "Pakistan"), -0.171)
+})
+
 test_that("EQ-5D-3L Poland TTO gives correct answer", {
   expect_equal(eq5d3l(c(MO=1,SC=1,UA=1,PD=1,AD=1), "TTO", "Poland"), 1)
   expect_equal(eq5d3l(c(MO=1,SC=1,UA=1,PD=1,AD=2), "TTO", "Poland"), 0.925)