+++ NEWS.md
@@ -1,6 +1,10 @@
 
 <!-- NEWS.md is generated from NEWS.Rmd. Please edit that file -->
 
+# eq5d (development version)
+
+- Pakistan EQ-5D-3L value set added.
+
 # eq5d 0.15.0 (2023-07-13)
 
 - Iranian and Slovenian EQ-5D-5L value sets added.